Fix index-out-of-bounds crash
With certain positions (usually those on the start of the row) this can
crash with java.lang.IndexOutOfBoundsException. The problem is that
we're checking the position with mRowForChildPosition.size() and then
computing all the child sizes *up to* the position (ie. not including
the position). This can lead to crashes as can be seen in
https://github.com/500px/greedo-layout-for-android/issues/20